modules/m2-video-metadata/requirements.txt
# M2 Video Metadata Service - Dependencies

# Web Framework
flask==2.3.3
gunicorn==21.2.0
werkzeug==2.3.7

# GCP Services
google-cloud-bigquery==3.11.4
google-cloud-storage==2.10.0
google-cloud-logging==3.5.0

# Video Processing
ffmpeg-python==0.2.0

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1

# Development
python-dotenv==1.0.0
